get-mcu-stubs: refactor mcu stubber
